+++ modules/canvaskit/tests/core.spec.js
@@ -764,7 +764,8 @@ describe('Core canvas behavior', () => {
 
         // rotate 10 degrees centered on 200, 200
         const m = CanvasKit.Matrix.rotated(Math.PI/18, 200, 200);
-        const rotated = CanvasKit.ImageFilter.MakeMatrixTransform(m, CanvasKit.FilterQuality.Medium, combined);
+        const filtering = { filter: CanvasKit.FilterMode.Linear };
+        const rotated = CanvasKit.ImageFilter.MakeMatrixTransform(m, filtering, combined);
         paint.setImageFilter(rotated);
 
         //canvas.rotate(10, 200, 200);