+++ modules/canvaskit/CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Breaking
  - Paint.getBlendMode() has been removed.
+ - Canvas.drawImageAtCurrentFrame() has been removed.
 
 ## [0.28.1] - 2021-06-28