+++ modules/canvaskit/CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Breaking
  - Paint.getBlendMode() has been removed.
  - Canvas.drawImageAtCurrentFrame() has been removed.
+ - FilterQuality enum removed -- pass FilterOptions | CubicResampler instead.
 
 ## [0.28.1] - 2021-06-28