+++ modules/canvaskit/CHANGELOG.md
@@ -13,7 +13,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Breaking
  - `Paint.getBlendMode()` has been removed.
  - `Canvas.drawImageAtCurrentFrame()` has been removed.
- - FilterQuality enum removed -- pass FilterOptions | CubicResampler instead.
+ - FilterQuality enum removed -- pass `FilterOptions` | `CubicResampler` instead.
+
+### Type Changes (index.d.ts)
+ - Replaced all `object` with actual types, including `AnimationMarker`.
 
 ## [0.28.1] - 2021-06-28