+++ modules/canvaskit/CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.29.0] - 2021-08-06
+
 ### Added
  - `Path.makeAsWinding` has been added to convert paths with an EvenOdd FillType to the
    equivalent area using the Winding FillType.